Backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart
